Remove `on_hook` blocker in main loop
This runs so frequently that the Pi blocks all other events while on the
hook. Adding a sleep doesn't seem to help. Relying on the `on_hook`
variable as a blocker _within_ the existing interrupt handlers is okay.

Removing this blocker allows the existing `sleep` to kick in, allowing
the Pi to process background events.